Rename l18n to i18n
